All threads appear to correctly communicate with each other using muteces and condition variables. Test function appears to work.

+++ Analytics.h
@@ -9,9 +9,9 @@
 
 time_t starting_time;
 int total_number_of_jobs;
-float turnaround_time;
-float CPU_time;
-float waiting_time;
+long turnaround_time;
+long CPU_time;
+long waiting_time;
 float throughput;
 time_t job_start_time;
